temp rename to fasttext_module
