+++ scripts/download_misc_dependencies.py
@@ -8,7 +8,7 @@ from typing import Any, List
 
 from stanfordnlp.utils.resources import download
 
-from cltkv1.embeddings.fasttext import (
+from cltkv1.embeddings.fasttext_module import (
     are_fasttext_models_downloaded,
     download_fasttext_models,
 )
@@ -71,10 +71,10 @@ def get_fasttext_models(force: bool = True, lang: str = "all"):
         common_crawl_models.append(lang)
     for lang in wiki_models:
         print("Going to download the fastText 'wiki' models ...")
-        download_fasttext_models(iso_code=lang, vector_type="wiki", force=force)
+        download_fasttext_models(iso_code=lang, training_set="wiki", force=force)
     for lang in common_crawl_models:
         print("Going to download the fastText 'common_crawl' models ...")
-        download_fasttext_models(iso_code=lang, vector_type="common_crawl", force=force)
+        download_fasttext_models(iso_code=lang, training_set="common_crawl", force=force)
 
 
 if __name__ == "__main__":
